refac: CartDisplay 컴포넌트 세부 분리
